requirements.txt
django>=3.1.12
psycopg2-binary==2.8.5
#celery==4.4.6
djangorestframework>=3.11.2
drf-yasg
django-model-utils==4.1.1
